fix: code review fixes — extract safeCompare, sanitize backupPath, unify paths, add input validation
- Extract duplicated safeCompare() to shared src/auth/utils.js
- Sanitize backupPath in undo endpoint to prevent path traversal
- Replace os.homedir() with resolveHomeDir()/resolveConfigDir() in optimizer and openclaw_config
- Add input validation (savings, meta) on POST /api/optimize/:action_id
- D05: only trigger on explicit thinkingDefault config, skip undefined
- D01: add comment clarifying intentional single-model recommendation
- Remove unused cacheWrite from diagnostics fallback
- Add OPENCLAW_STATE_DIR override in integration tests
