+++ src/auth/middleware.js
@@ -3,23 +3,9 @@
  */
 const path = require('path');
 const fs = require('fs');
-const crypto = require('crypto');
 const { SECRET_KEY } = require('../config');
 const { hasSession, generateSessionToken, addSession } = require('./sessions');
-
-// Timing-safe key comparison to prevent timing attacks
-function safeCompare(a, b) {
-    if (typeof a !== 'string' || typeof b !== 'string') return false;
-    const bufA = Buffer.from(a);
-    const bufB = Buffer.from(b);
-    if (bufA.length !== bufB.length) {
-        // Hash both to constant-time compare even when lengths differ
-        const hashA = crypto.createHash('sha256').update(bufA).digest();
-        const hashB = crypto.createHash('sha256').update(bufB).digest();
-        return crypto.timingSafeEqual(hashA, hashB);
-    }
-    return crypto.timingSafeEqual(bufA, bufB);
-}
+const { safeCompare } = require('./utils');
 
 // Cache login page HTML at startup
 const LOGIN_PAGE = fs.readFileSync(path.join(__dirname, 'login.html'), 'utf8');
