+++ src/auth/routes.js
@@ -2,7 +2,6 @@
  * Auth routes — POST /api/auth, POST /api/logout
  */
 const router = require('express').Router();
-const crypto = require('crypto');
 const { SECRET_KEY } = require('../config');
 const {
     generateSessionToken,
@@ -11,18 +10,7 @@ const {
     checkAuthRateLimit,
     resetAuthAttempts,
 } = require('./sessions');
-
-function safeCompare(a, b) {
-    if (typeof a !== 'string' || typeof b !== 'string') return false;
-    const bufA = Buffer.from(a);
-    const bufB = Buffer.from(b);
-    if (bufA.length !== bufB.length) {
-        const hashA = crypto.createHash('sha256').update(bufA).digest();
-        const hashB = crypto.createHash('sha256').update(bufB).digest();
-        return crypto.timingSafeEqual(hashA, hashB);
-    }
-    return crypto.timingSafeEqual(bufA, bufB);
-}
+const { safeCompare } = require('./utils');
 
 // POST /api/auth — Login
 router.post('/api/auth', (req, res) => {
